requirements.txt
customtkinter
pycryptodome
numpy
